Switch from using Asim's sysreg defines to the AP_* registers defined in the CSR yaml. A nice side affect of this is you can now access the system registers of the local core using "csr AP_*".
git-svn-id: http://casvn/svn/octeon/bdk/branches/thunder@123378 c83883e8-9f28-0410-baf1-f2025811e16a

+++ libbdk-arch/bdk-csr.h
@@ -61,8 +61,8 @@ extern int __bdk_csr_lookup_index(const char *name, int params[]);
 extern int bdk_csr_get_name(const char *last_name, char *buffer);
 struct bdk_readline_tab;
 extern struct bdk_readline_tab *__bdk_csr_get_tab_complete() BDK_WEAK;
-extern uint64_t bdk_sysreg_read(int node, int core, int regnum);
-extern void bdk_sysreg_write(int node, int core, int regnum, uint64_t value);
+extern uint64_t bdk_sysreg_read(int node, int core, uint64_t regnum);
+extern void bdk_sysreg_write(int node, int core, uint64_t regnum, uint64_t value);
 
 #ifndef BDK_BUILD_HOST
